// src/interfaces/vax_malaysia.ts
interface VaxMalaysia {
    date: string;
    daily_partial: string;
    daily_full: string;
    daily_booster: string;
    daily: string;
    daily_partial_adol: string;
    daily_full_adol: string;
    daily_partial_child: string;
    daily_full_child: string;
    cumul_partial: string;
    cumul_full: string;
    cumul_booster: string;
    cumul: string;
    cumul_partial_adol: string;
    cumul_full_adol: string;
    cumul_partial_child: string;
    cumul_full_child: string;
    pfizer1: string;
    pfizer2: string;
    pfizer3: string;
    sinovac1: string;
    sinovac2: string;
    sinovac3: string;
    astra1: string;
    astra2: string;
    astra3: string;
    sinopharm1: string;
    sinopharm2: string;
    sinopharm3: string;
    cansino: string;
    cansino3: string;
    pending1: string;
    pending2: string;
    pending3: string;
}

export default VaxMalaysia;