// src/interfaces/registration.ts
interface Registration {
    date: string;
    state: string;
    total: string;
    phase2: string;
    mysj: string;
    call: string;
    web: string;
    children: string;
    children_solo: string;
    adolescents: string;
    elderly: string;
    comorb: string;
    oku: string;
}

export default Registration;